re-org InterfaceAdapter type defs & import appropriately (interface-adapter)
